lib: Split out from Listing the component for the content of expandable list items
The purpose of this split is to reuse this component is ListingTable
while we are slowly porting all our tables to PF4 tables.
The old implementation of Listing component will continue to use the
ListingPanel as well.
